+++ src/threads.js
@@ -62,7 +62,7 @@ StageMorph, SpriteMorph, StagePrompterMorph, Note, modules, isString, copy,
 isNil, WatcherMorph, List, ListWatcherMorph, alert, console, TableMorph, Color,
 TableFrameMorph, ColorSlotMorph, isSnapObject, Map*/
 
-modules.threads = '2019-March-06';
+modules.threads = '2019-March-10';
 
 var ThreadManager;
 var Process;
@@ -2251,12 +2251,16 @@ Process.prototype.doStopAllSounds = function () {
 
 // Process audio input (interpolated)
 
-Process.prototype.reportAudio = function (choice) { // +++
+Process.prototype.reportAudio = function (choice) {
     var stage = this.blockReceiver().parentThatIsA(StageMorph);
     if (stage.microphone.isOn()) {
         switch (this.inputOption(choice)) {
         case 'volume':
             return stage.microphone.volume;
+        case 'pitch':
+            return stage.microphone.pitch;
+        case 'note':
+            return stage.microphone.note;
         case 'signals':
             return new List(stage.microphone.signals);
         case 'frequencies':